Fix data type mismatches to silent compiler warnings.

+++ src/PhysicalSystems/Heisenberg2D.hpp
@@ -24,8 +24,8 @@ public :
 
 private :
 
-  int Size;
-  long LatticeSize;
+  unsigned int Size;
+  unsigned int LatticeSize;
 
   struct SpinDirection {
     double x;
